+++ assignment1/main/SimulPar.java
@@ -1,11 +1,11 @@
 package assignment1.main;
 
 public class SimulPar {
-    public static final short NG = 3; // number of games, default = 3
-    public static final short NT = 6; // number of trials, default = 4
-    public static final short KT = 4; // knockout treshold, default = 4
-    public static final short NC = 5; // number of contestants per team, default = 5
-    public static final short NP = 3; // number of active player per team, default = 3
-    public static final short MINS = 10; // min strength of contestants, default = 6
-    public static final short MAXS = 15; // max strength of contestants, default = 10
+    public static final int NG = 3; // number of games, default = 3
+    public static final int NT = 6; // number of trials, default = 4
+    public static final int KT = 4; // knockout treshold, default = 4
+    public static final int NC = 5; // number of contestants per team, default = 5
+    public static final int NP = 3; // number of active player per team, default = 3
+    public static final int MINS = 10; // min strength of contestants, default = 6
+    public static final int MAXS = 15; // max strength of contestants, default = 10
 }
